refactor: 多くのメッセージを寄せられるように、クラスの階層を一つ上に上げる。
