+++ src/components/main/tree/DiagramsTreePane.vue
@@ -39,7 +39,7 @@ import DiagramRightClickMenu from './DiagramRightClickMenu.vue'
 import Folder from './Folder'
 import TreeWrapper from './TreeWrapper'
 import FolderTreeFactory from './FolderTreeFactory'
-import Prompts from './Prompts'
+import Prompts from '@/components/main/Prompts'
 import DiagramExportService from '@/application/service/diagram/export/DiagramExportService'
 import Diagram from '@/domain/diagram/Diagram'
 import Product from '@/domain/product/Product'
